post,comment,test-support,profile-tests: three more finds from the CI matrix
- post/comment: LIMIT bound as i64 against CQL int32 — every paged list
  failed serialization (the soak's social-graph #17, same class, never swept
  to siblings; these paths had no live traffic).
- test-support: migration loaders were non-recursive — moderation's split
  migrations/{postgres,scylla} loaded NOTHING and its suite ran against an
  empty schema (relation "cases" does not exist).
- profile harness: random_handle() took the first 12 hex of a UUID v7 — i.e.
  the unix-millis timestamp — so scenarios starting in the same millisecond
  generated IDENTICAL handles: cross-test LWT conflicts produced "fresh
  handle already taken" and the zero-winner claim race. v4 random tail now;
  all 3 scenarios green locally against real Scylla.

+++ crates/services/comment/src/infrastructure/persistence/scylla_comment_repository.rs
@@ -404,7 +404,11 @@ impl CommentRepository for ScyllaCommentRepository {
         limit:      i32,
         page_token: Option<&str>,
     ) -> Result<(Vec<CommentSummary>, Option<String>), CommentError> {
-        let limit = limit.clamp(1, 100) as i64;
+        // i32, NOT i64: CQL `LIMIT ?` type-checks against int32 and the i64 bind
+        // fails serialization — the soak's social-graph finding (#17), same class,
+        // never swept here (these list paths had no live traffic until the suites
+        // ran in CI).
+        let limit = limit.clamp(1, 100);
         let token = decode_page_token(page_token)?;
 
         let rows: Vec<CommentFeedRow> = if let Some(ref tok) = token {
